get user location

+++ app/src/main/java/com/example/myweatherapp/service/WeatherService.kt
@@ -8,8 +8,8 @@ import retrofit2.http.Query
 interface WeatherService {
     @GET("forecast")
     suspend fun forecast(
-        @Query("latitude") latitude: Float,
-        @Query("longitude") longitude: Float,
+        @Query("latitude") latitude: Double,
+        @Query("longitude") longitude: Double,
         @Query("hourly") hourly: String = "apparent_temperature",
         @Query("daily") daily: String = "weathercode,temperature_2m_max,temperature_2m_min,apparent_temperature_max,apparent_temperature_min,precipitation_sum,windspeed_10m_max",
         @Query("current_weather") current_weather: String = "true",
